+++ src/main/java/com/unzer/payment/communication/UnzerHttpRequest.java
@@ -16,6 +16,7 @@
 package com.unzer.payment.communication;
 
 import java.net.URI;
+import java.net.URISyntaxException;
 
 /**
  * Abstraction for any http-request executed by the
@@ -26,7 +27,7 @@ public interface UnzerHttpRequest {
 
     void addHeader(String header, String value);
 
-    URI getURI();
+    URI getURI() throws URISyntaxException;
 
     void setContent(String content, String encoding);