+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.1][3.1.1]
+
+This release updates dependencies with security issues
+
+### Changed
+
+* Upgraded Apache HttpClient from version 4 to version 5.
+* Made `HttpCommunicationException` unchecked exception.
+* Updated `jackson-core` dependency.
+
 ## [3.1.0][3.1.0]
 
 This release introduces Unzer PayPal Express in Java SDK.
@@ -314,6 +324,8 @@ This release brings Unzer Paylater Invoice payment type support to Java SDK.
     *   cancelAuthorization
 *   Remove deprecated classes
     *   RestCommunication
+
+[3.1.1]: http://github.com/unzerdev/java-sdk/compare/3.1.0..3.1.1
 
 [3.1.0]: http://github.com/unzerdev/java-sdk/compare/3.0.0..3.1.0